Set up Semantic UI, began to rebuild interface
See semantic UI Docs to see how the HTML templating is being done.  ALso, it appears when a file is deleted, if that file exists elsewhere (i.e. same file is uploaded under different names) the file is removed so that the one left over no longer has a file to access